blur-effect: Use int for opacity override
The opacity override can be -1 if no override is set, so use the
appropriate datatype here.

https://gitlab.gnome.org/GNOME/gnome-shell/merge_requests/991

+++ src/shell-blur-effect.c
@@ -158,7 +158,7 @@ struct _ShellBlurEffect
   ClutterEffect parent_instance;
 
   ClutterActor *actor;
-  uint8_t old_opacity_override;
+  int old_opacity_override;
 
   BlurData blur[2];
